requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
google-api-core==1.26.3
google-auth==1.28.0
google-cloud-core==1.6.0
google-cloud-translate==2.0.1
googleapis-common-protos==1.53.0
grpcio==1.36.1
idna==2.10
packaging==20.9
protobuf==3.15.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
urllib3==1.26.4
